build: Make MSRV explicit
Add `rust-version` to `Cargo.toml`. Verify this MSRV by adding a CI check that builds using that rust compiler version.

The current MSRV is 1.61. Trying to build with 1.60 results in:
```
error: package `syn v2.0.90` cannot be built because it requires rustc 1.61 or newer, while the currently active rustc version is 1.60.0
```